data-server/pyproject.toml
[project]
name = "edu-max-data"
version = "0.1.0"
description = "edu max data"
readme = "README.md"
requires-python = ">=3.11"
dependencies=[
    "aiofiles>=25.1.0",
    "fastapi==0.115.12",
    "huggingface-hub==0.30.1",
    "langchain==0.3.14",
    "langchain-community==0.3.1",
    "langchain-core==0.3.62",
    "langchain-huggingface==0.1.0",
    "langchain-weaviate==0.0.4",
    "pandas==2.2.3",
    "python-multipart==0.0.20",
    "rich==13.9.4",
    "uvicorn==0.34.0",
    "uvloop>=0.22.1",
    "watchfiles==1.0.4",
    "weaviate-client==4.14.4",
]

[tool.uv]
python-preference = "managed"
managed = true
